short -> int
git-svn-id: http://svn.code.sf.net/p/dnsjava/code/trunk@1050 c76caeb1-94fd-44dd-870f-0c9d92034fc1

+++ lookup.java
@@ -38,7 +38,7 @@ main(String [] args) throws Exception {
 		start = 2;
 	}
 	for (int i = start; i < args.length; i++) {
-		Lookup l = new Lookup(args[i], (short) type);
+		Lookup l = new Lookup(args[i], type);
 		l.run();
 		printAnswer(args[i], l);
 	}
